Adding current standing of P4 of lab assignment
Code is mostly correct, however it is not correctly outputting one of the independent paths.
Current output:
0->1->2->3
0->1->3
0->2->3

Expected:
0->1->2->3
0->1->3
0->2

It is double counting the edge between nodes 2 and 3. Not 100% sure how to deal with it so I am sharing my work here in case you guys want to take a crack at it.